+++ app/src/main/java/shopify/user/shopifytest/core/model/product/Image.java
@@ -9,10 +9,10 @@ public class Image {
 
     @SerializedName("id")
     @Expose
-    private Integer id;
+    private Long id;
     @SerializedName("product_id")
     @Expose
-    private Integer productId;
+    private Long productId;
     @SerializedName("position")
     @Expose
     private Integer position;
@@ -41,19 +41,19 @@ public class Image {
     @Expose
     private String adminGraphqlApiId;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(Integer productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
