+++ app/src/main/java/shopify/user/shopifytest/core/model/product/Option.java
@@ -9,10 +9,10 @@ public class Option {
 
     @SerializedName("id")
     @Expose
-    private Integer id;
+    private Long id;
     @SerializedName("product_id")
     @Expose
-    private Integer productId;
+    private Long productId;
     @SerializedName("name")
     @Expose
     private String name;
@@ -23,19 +23,19 @@ public class Option {
     @Expose
     private List<String> values = null;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(Integer productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
